@@ -4,7 +4,7 @@
   "description": "A manifesto project website.",
   "main": "src/index.js",
   "scripts": {
-    "start": "react-scripts start",
+    "start": "export NODE_OPTIONS=--openssl-legacy-provider && react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
@@ -18,11 +18,10 @@
   "author": "Your Name",
   "license": "MIT",
   "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-scripts": "4.0.3"
   },
-  "devDependencies": {},
   "homepage": "http://rngnnn.github.io/manifesto-project",
   "browserslist": {
     "production": [
