@@ -1,19 +1,17 @@
+// src/components/Footer.jsx
 import React from 'react';
-import '../styles/Footer.css';
+import './Footer.css';
 
-function Footer() {
+const Footer = () => {
   return (
     <footer className="footer">
-      <div className="footer-content">
-        <p>&copy; {new Date().getFullYear()} .</p>
-        <div className="footer-links">
-          <a href="#about">About</a>
-          <a href="#contact">Contact</a>
-          <a href="#privacy">Privacy Policy</a>
-        </div>
-      </div>
+      <p>© {new Date().getFullYear()} Your Name. All rights reserved.  
+        <br />
+        <a href="https://github.com/yourprofile" target="_blank" rel="noreferrer">GitHub</a> |  
+        <a href="mailto:you@example.com"> Contact</a>
+      </p>
     </footer>
   );
-}
+};
 
 export default Footer;
