@@ -1,20 +1,19 @@
+// src/components/Header.jsx
 import React from 'react';
-import '../styles/Header.css'; // CSS dosyasını dahil ediyoruz
+import './Header.css';
 
-function Header() {
+const Header = () => {
   return (
     <header className="header">
-      <div className="logo"></div>
-      <nav className="navigation">
-        <ul>
-          <li><a href="#home">Home</a></li>
-          <li><a href="#about">About</a></li>
-          <li><a href="#services">Services</a></li>
-          <li><a href="#contact">Contact</a></li>
-        </ul>
+      <a href="#" className="header-logo">MyLogo</a>
+      <nav className="header-nav">
+        <a href="#home">Home</a>
+        <a href="#about">About</a>
+        <a href="#projects">Projects</a>
+        <a href="#contact">Contact</a>
       </nav>
     </header>
   );
-}
+};
 
 export default Header;
